refactor: separate artifact generation logic into StarknetArtifactGenerator class and improve template handling
