+++ starknet/scripts/Templates.ts
